pyproject.toml
[project]
name = "njx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "absl-py>=2.1.0",
    "arbor>=0.10.0",
    "furo>=2024.8.6",
    "jupyter-book>=0.6.5",
    "jupyter-sphinx>=0.5.3",
    "jupytext>=1.16.4",
    "matplotlib>=3.9.2",
    "myst-nb>=1.1.2",
    "pytest>=8.3.3",
    "ruff>=0.7.0",
    "sphinx-autodoc-typehints>=2.5.0",
    "sphinx-copybutton>=0.5.2",
    "sphinx-remove-toctrees>=1.0.0.post1",
    "sphinx>=8.1.3",
    "tree-math>=0.2.1",
]
